pyproject.toml
[tool.poetry]
name = "raspberry-pi-weather-display"
version = "0.1.0"
description = "E-Ink weather dashboard for Raspberry Pi Zero 2 W"
authors = ["Steve Nims <sjnims@gmail.com>"]
license = "MIT"
keywords = ["raspberry-pi", "e-ink", "weather", "dashboard", "iot"]
readme = "README.md"
homepage = "https://github.com/sjnims/raspberry-pi-weather-display"
repository = "https://github.com/sjnims/raspberry-pi-weather-display"
documentation = "https://github.com/sjnims/raspberry-pi-weather-display#readme"
packages = [
  { include = "rpiweather", from = "src" },
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

include = [
  { path = "deploy/scripts/install.sh", format = "sdist" },
  { path = "deploy/weather-display.service", format = "sdist" },
]

[tool.poetry.scripts]
weather = "rpiweather.cli:app"

[tool.poetry.dependencies]
python = "^3.11"
jinja2 = "^3.1"
PyYAML = "^6.0"
requests = "^2.31"
typer = { extras = ["all"], version = "^0.15.2" }
pydantic = ">=2"
pillow = "^11.2.1"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1"
pytest = "^8.0"
pytest-asyncio = "^0.23"
ruff = "^0.11.8"
types-jinja2 = "^2.11"
types-pyyaml = "^6.0"
types-requests = "^2.32"
pre-commit = "^4.2.0"
python-dotenv = "^1.1.0"
coverage = "^7.8.0"
pytest-cov = "^6.1.1"

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "I", "UP", "SIM", "B", "C4", "PIE", "RUF"]
lint.ignore = ["E501"]  # optionally ignore long lines if needed
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pyright]
reportMissingImports = true
reportMissingTypeStubs = false
typeCheckingMode = "strict"

[tool.coverage.run]
source = ["rpiweather"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"